Deduplicate the `IndexFileBuffer` hierarchy into separate, pakage-private interface `IndexFileBuffer` and classes `IndexFileBuffer`, `MemoryMappedFileBuffer` and `RandomAccessFileBuffer`. Refactor the code to use the new classes.
